Move more code out of cellRenderer #131808
