+++ src/vs/workbench/contrib/notebook/browser/view/renderers/cellRenderer.ts
@@ -7,19 +7,13 @@ import { PixelRatio } from 'vs/base/browser/browser';
 import * as DOM from 'vs/base/browser/dom';
 import { FastDomNode } from 'vs/base/browser/fastDomNode';
 import { IListRenderer, IListVirtualDelegate } from 'vs/base/browser/ui/list/list';
-import { Color } from 'vs/base/common/color';
 import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
-import * as platform from 'vs/base/common/platform';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { CodeEditorWidget } from 'vs/editor/browser/widget/codeEditorWidget';
-import { EditorOption, IEditorOptions } from 'vs/editor/common/config/editorOptions';
+import { IEditorOptions } from 'vs/editor/common/config/editorOptions';
 import { BareFontInfo } from 'vs/editor/common/config/fontInfo';
-import { Range } from 'vs/editor/common/core/range';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
-import * as languages from 'vs/editor/common/languages';
 import { PLAINTEXT_LANGUAGE_ID } from 'vs/editor/common/languages/modesRegistry';
-import { tokenizeLineToHTML } from 'vs/editor/common/languages/textToHtmlTokenizer';
-import { ITextModel } from 'vs/editor/common/model';
 import { localize } from 'vs/nls';
 import { IMenuService } from 'vs/platform/actions/common/actions';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
@@ -32,9 +26,11 @@ import { INotificationService } from 'vs/platform/notification/common/notificati
 import { ICellViewModel, INotebookEditorDelegate } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { CellContextKeyManager } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellContextKeys';
 import { CellDecorations } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellDecorations';
-import { CellDragAndDropController } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellDnd';
+import { CellDragAndDropController, CellDragAndDropPart } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellDnd';
+import { CodeCellDragImageRenderer } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellDragRenderer';
 import { CellEditorOptions } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellEditorOptions';
 import { CellExecutionPart } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellExecution';
+import { CellFocusPart } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellFocus';
 import { CellFocusIndicator } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellFocusIndicator';
 import { CellProgressBar } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellProgressBar';
 import { CellEditorStatusBar } from 'vs/workbench/contrib/notebook/browser/view/cellParts/cellStatusPart';
@@ -109,16 +105,7 @@ abstract class AbstractCellRenderer {
 		this.dndController = undefined;
 	}
 
-	protected commonRenderTemplate(templateData: BaseCellRenderTemplate): void {
-		templateData.templateDisposables.add(DOM.addDisposableListener(templateData.container, DOM.EventType.FOCUS, () => {
-			if (templateData.currentRenderedCell) {
-				this.notebookEditor.focusElement(templateData.currentRenderedCell);
-			}
-		}, true));
-	}
-
 	protected commonRenderElement(element: ICellViewModel, templateData: BaseCellRenderTemplate): void {
-		this.dndController?.renderElement(element, templateData);
 		templateData.elementDisposables.add(new CellDecorations(templateData.rootContainer, templateData.decorationContainer, element));
 		templateData.elementDisposables.add(templateData.instantiationService.createInstance(CellContextKeyManager, this.notebookEditor, element));
 	}
@@ -191,6 +178,8 @@ export class MarkupCellRenderer extends AbstractCellRenderer implements IListRen
 			focusIndicator,
 			foldedCellHint,
 			templateDisposables.add(new CollapsedCellInput(this.notebookEditor, cellInputCollapsedContainer)),
+			templateDisposables.add(new CellFocusPart(container, this.notebookEditor)),
+			templateDisposables.add(new CellDragAndDropPart()),
 		];
 
 		const templateData: MarkdownCellRenderTemplate = {
@@ -209,8 +198,6 @@ export class MarkupCellRenderer extends AbstractCellRenderer implements IListRen
 			cellParts,
 			toJSON: () => { return {}; }
 		};
-
-		this.commonRenderTemplate(templateData);
 
 		return templateData;
 	}
@@ -240,116 +227,6 @@ export class MarkupCellRenderer extends AbstractCellRenderer implements IListRen
 
 	disposeElement(_element: ICellViewModel, _index: number, templateData: MarkdownCellRenderTemplate): void {
 		templateData.elementDisposables.clear();
-	}
-}
-
-class EditorTextRenderer {
-
-	private static _ttPolicy = window.trustedTypes?.createPolicy('cellRendererEditorText', {
-		createHTML(input) { return input; }
-	});
-
-	getRichText(editor: ICodeEditor, modelRange: Range): HTMLElement | null {
-		const model = editor.getModel();
-		if (!model) {
-			return null;
-		}
-
-		const colorMap = this.getDefaultColorMap();
-		const fontInfo = editor.getOptions().get(EditorOption.fontInfo);
-		const fontFamilyVar = '--notebook-editor-font-family';
-		const fontSizeVar = '--notebook-editor-font-size';
-		const fontWeightVar = '--notebook-editor-font-weight';
-
-		const style = ``
-			+ `color: ${colorMap[languages.ColorId.DefaultForeground]};`
-			+ `background-color: ${colorMap[languages.ColorId.DefaultBackground]};`
-			+ `font-family: var(${fontFamilyVar});`
-			+ `font-weight: var(${fontWeightVar});`
-			+ `font-size: var(${fontSizeVar});`
-			+ `line-height: ${fontInfo.lineHeight}px;`
-			+ `white-space: pre;`;
-
-		const element = DOM.$('div', { style });
-
-		const fontSize = fontInfo.fontSize;
-		const fontWeight = fontInfo.fontWeight;
-		element.style.setProperty(fontFamilyVar, fontInfo.fontFamily);
-		element.style.setProperty(fontSizeVar, `${fontSize}px`);
-		element.style.setProperty(fontWeightVar, fontWeight);
-
-		const linesHtml = this.getRichTextLinesAsHtml(model, modelRange, colorMap);
-		element.innerHTML = linesHtml as string;
-		return element;
-	}
-
-	private getRichTextLinesAsHtml(model: ITextModel, modelRange: Range, colorMap: string[]): string | TrustedHTML {
-		const startLineNumber = modelRange.startLineNumber;
-		const startColumn = modelRange.startColumn;
-		const endLineNumber = modelRange.endLineNumber;
-		const endColumn = modelRange.endColumn;
-
-		const tabSize = model.getOptions().tabSize;
-
-		let result = '';
-
-		for (let lineNumber = startLineNumber; lineNumber <= endLineNumber; lineNumber++) {
-			const lineTokens = model.getLineTokens(lineNumber);
-			const lineContent = lineTokens.getLineContent();
-			const startOffset = (lineNumber === startLineNumber ? startColumn - 1 : 0);
-			const endOffset = (lineNumber === endLineNumber ? endColumn - 1 : lineContent.length);
-
-			if (lineContent === '') {
-				result += '<br>';
-			} else {
-				result += tokenizeLineToHTML(lineContent, lineTokens.inflate(), colorMap, startOffset, endOffset, tabSize, platform.isWindows);
-			}
-		}
-
-		return EditorTextRenderer._ttPolicy?.createHTML(result) ?? result;
-	}
-
-	private getDefaultColorMap(): string[] {
-		const colorMap = languages.TokenizationRegistry.getColorMap();
-		const result: string[] = ['#000000'];
-		if (colorMap) {
-			for (let i = 1, len = colorMap.length; i < len; i++) {
-				result[i] = Color.Format.CSS.formatHex(colorMap[i]);
-			}
-		}
-		return result;
-	}
-}
-
-class CodeCellDragImageRenderer {
-	getDragImage(templateData: BaseCellRenderTemplate, editor: ICodeEditor, type: 'code' | 'markdown'): HTMLElement {
-		let dragImage = this.getDragImageImpl(templateData, editor, type);
-		if (!dragImage) {
-			// TODO@roblourens I don't think this can happen
-			dragImage = document.createElement('div');
-			dragImage.textContent = '1 cell';
-		}
-
-		return dragImage;
-	}
-
-	private getDragImageImpl(templateData: BaseCellRenderTemplate, editor: ICodeEditor, type: 'code' | 'markdown'): HTMLElement | null {
-		const dragImageContainer = templateData.container.cloneNode(true) as HTMLElement;
-		dragImageContainer.classList.forEach(c => dragImageContainer.classList.remove(c));
-		dragImageContainer.classList.add('cell-drag-image', 'monaco-list-row', 'focused', `${type}-cell-row`);
-
-		const editorContainer: HTMLElement | null = dragImageContainer.querySelector('.cell-editor-container');
-		if (!editorContainer) {
-			return null;
-		}
-
-		const richEditorText = new EditorTextRenderer().getRichText(editor, new Range(1, 1, 1, 1000));
-		if (!richEditorText) {
-			return null;
-		}
-		DOM.reset(editorContainer, richEditorText);
-
-		return dragImageContainer;
 	}
 }
 
@@ -453,6 +330,8 @@ export class CodeCellRenderer extends AbstractCellRenderer implements IListRende
 			templateDisposables.add(new CellExecutionPart(this.notebookEditor, executionOrderLabel)),
 			templateDisposables.add(this.instantiationService.createInstance(CollapsedCellOutput, this.notebookEditor, cellOutputCollapsedContainer)),
 			templateDisposables.add(new CollapsedCellInput(this.notebookEditor, cellInputCollapsedContainer)),
+			templateDisposables.add(new CellFocusPart(container, this.notebookEditor)),
+			templateDisposables.add(new CellDragAndDropPart()),
 		];
 
 		const templateData: CodeCellRenderTemplate = {
@@ -485,8 +364,6 @@ export class CodeCellRenderer extends AbstractCellRenderer implements IListRende
 				this.notebookEditor.focusNotebookCell(templateData.currentRenderedCell, 'output');
 			}
 		}));
-
-		this.commonRenderTemplate(templateData);
 
 		return templateData;
 	}
